feat: add dependency injection extensions
