+++ src/SocialMediaDashboard.WebAPI/Extensions/WebServiceCollectionExtension.cs
@@ -1,0 +1,52 @@
+﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.IdentityModel.Tokens;
+using SocialMediaDashboard.Common.Helpers;
+using System.Text;
+
+namespace SocialMediaDashboard.WebAPI.Extensions
+{
+    /// <summary>
+    /// Service collection for Web project.
+    /// </summary>
+    public static class WebServiceCollectionExtension
+    {
+        /// <summary>
+        /// Dependency injection.
+        /// </summary>
+        /// <param name="services">Service collection.</param>
+        /// <param name="configuration">Configuration.</param>
+        /// <returns>Service collection.</returns>
+        public static IServiceCollection AddWeb(this IServiceCollection services, IConfiguration configuration)
+        {
+            var appSettingsSection = configuration.GetSection("AppSettings");
+            services.Configure<ApplicationSettings>(appSettingsSection);
+
+            var appSettings = appSettingsSection.Get<ApplicationSettings>();
+            var key = Encoding.ASCII.GetBytes(appSettings.Secret);
+            services.AddAuthentication(x =>
+            {
+                x.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
+                x.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+            })
+            .AddJwtBearer(x =>
+            {
+                x.RequireHttpsMetadata = false;
+                x.SaveToken = true;
+                x.TokenValidationParameters = new TokenValidationParameters
+                {
+                    ValidateIssuerSigningKey = true,
+                    IssuerSigningKey = new SymmetricSecurityKey(key),
+                    ValidateIssuer = false,
+                    ValidateAudience = false
+                };
+            });
+
+            services.AddCors();
+            services.AddControllers();
+
+            return services;
+        }
+    }
+}
