+++ src/SocialMediaDashboard.WebAPI/Properties/launchSettings.json
@@ -20,7 +20,7 @@
     "SocialMediaDashboard.WebAPI": {
       "commandName": "Project",
       "launchBrowser": true,
-      "launchUrl": "weatherforecast",
+      "launchUrl": "",
       "applicationUrl": "https://localhost:5001;http://localhost:5000",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
